Split progress render tests
